backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==1.26.4
sse-starlette==2.1.0
